Move shared controller methods to traits.
This will make it easy to extract them into shared packages when we get
a chance.
